add main.rs; fix broken rustdoc
